+++ Basic/Jenny/209Project/209Project/energyCalculations.h
@@ -14,8 +14,8 @@
 
 uint16_t calculateVrms(uint16_t* samples, uint8_t numSamples);
 uint16_t calculateIrms(uint16_t* samples, uint8_t numSamples);
-uint16_t calculateIpk(uint16_t Irms);
-uint16_t calculatePower(uint16_t Vrms, uint16_t Irms);
+float calculateIpk(float Irms);
+float calculatePower(float Vrms, float Irms);
 
 
 #endif /* ENERGYCALCULATIONS_H_ */
